requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.6
django-appconf==1.0.5
django-imagekit==4.1.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
easy-thumbnails==2.8.5
pilkit==2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
whitenoise==6.3.0
